Struct lock_keeper_postgres::api::PostgresDB

pub struct PostgresDB {
    config: Arc<Config>,
    connection_pool: PgPool,
}

Fields
config: Arc<Config>connection_pool: PgPool

PgPool is already implemented in terms of an Arc. No need to wrap it.

Implementations

impl PostgresDB

pub async fn connect(config: Config) -> Result<Self, PostgresError>

pub fn db_name(&self) -> &str

pub(crate) async fn create_audit_event_impl( &self, request_id: Uuid, account_id: AccountId, key_id: &Option<KeyId>, action: ClientAction, status: EventStatus ) -> Result<(), PostgresError>

async fn find_audit_events_impl( &self, account_id: AccountId, event_type: EventType, options: AuditEventOptions ) -> Result<Vec<AuditEvent>, PostgresError>

Create a dynamic query to fetch audit events specified by the caller.

pub(crate) async fn add_secret_impl( &self, secret: StoredSecret ) -> Result<(), PostgresError>

pub(crate) async fn get_secret_impl( &self, account_id: AccountId, key_id: &KeyId, filter: SecretFilter ) -> Result<StoredSecret, PostgresError>

This function verifies the user_id and key type matches. Otherwise will return a IncorrectKeyMetadata error.

pub(crate) async fn get_server_encrypted_blob_impl( &self, account_id: AccountId, key_id: &KeyId ) -> Result<StoredSecret, PostgresError>

This function provides a simplified query for server encrypted blobs.

pub(crate) async fn delete_secret_impl( &self, account_id: AccountId, key_id: &KeyId ) -> Result<(), PostgresError>

This function verifies the account_id and key_id match. Otherwise will return a NoEntry error.

pub(crate) async fn create_account_impl( &self, user_id: &UserId, account_name: &AccountName, server_registration: &ServerRegistration<OpaqueCipherSuite> ) -> Result<Account, PostgresError>

pub(crate) async fn find_account_by_name_impl( &self, account_name: &AccountName ) -> Result<Option<Account>, PostgresError>

pub(crate) async fn find_account_impl( &self, account_id: AccountId ) -> Result<Option<Account>, PostgresError>

pub(crate) async fn delete_account_impl( &self, account_id: AccountId ) -> Result<(), PostgresError>

pub(crate) async fn set_storage_key_impl( &self, account_id: AccountId, storage_key: Encrypted<StorageKey> ) -> Result<(), PostgresError>

pub(crate) async fn user_id_exists_impl( &self, user_id: &UserId ) -> Result<bool, PostgresError>